refactor: api actions moved to the slices
